parcelhubPOS/requirements.txt.txt
Django==1.11.4
django-addresses==0.1
django-backup==1.0.1
django-countries==4.6.1
django-crispy-forms==1.6.
django-datatable==0.3.1
django-dbbackup==3.2.0
django-excel==0.0.9
django-extensions==1.9.1
django-filter==1.0.4
django-iban==0.3.1
django-import-export==0.5
django-invoice==0.3.2
django-model-utils==3.0.0
django-money==0.11.4
django-simple-menu==1.2.1
django-tables2==1.10.0
djangorestframework==3.6.
et-xmlfile==1.0.1
inflect==0.2.5
jdcal==1.3
jsonfield==2.0.2
lml==0.0.1
mysql-connector-python==2
mysqlclient==1.3.4
num2words==0.5.5
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
pbr==3.1.1
pep8==1.7.0
Pillow==4.3.0
psycopg2==2.7.3.1
py-moneyed==0.7.0
pyexcel==0.5.3
pyexcel-io==0.5.1
pyexcel-webio==0.1.2
pyexcel-xls==0.5.0
pyexcel-xlsx==0.5.0.1
pyflakes==1.6.0
python-memcached==1.58
python-stdnum==1.7
pytz==2017.2
PyYAML==3.12
q==2.6
reportlab==3.4.0
six==1.10.0
stevedore==1.25.0
tablib==0.11.5
texttable==0.9.1
unicodecsv==0.14.1
vatnumber==1.2
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
xlrd==1.1.0
xlwt==1.3.0
